+++ lib/test_prof/recipes/rspec/any_fixture.rb
@@ -1,18 +1,11 @@
 # frozen_string_literal: true
 
 require "test_prof/any_fixture"
+require "test_prof/recipes/rspec/before_all"
 
 RSpec.shared_context "any_fixture:clean", with_clean_fixture: true do
-  before do
-    raise("Cannot use clean context without a transaction!") unless
-      open_transaction?
-
+  before_all do
     TestProf::AnyFixture.clean
-  end
-
-  def open_transaction?
-    pool = ActiveRecord::Base.connection_pool
-    pool.active_connection? && pool.connection.open_transactions > 0
   end
 end